% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_application.R
\name{shiny_application}
\alias{shiny_application}
\title{Shiny application}
\usage{
shiny_application(host = "0.0.0.0", port = 3838)
}
\arguments{
\item{host}{The IPv4 address that the application should listen on.}

\item{port}{The TCP port that the application should listen on.}
}
\description{
This function allows to launch the shiny app.
}
\examples{
shiny_application()
}
